requirements.txt
# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Transformers
transformers>=4.30.0
tokenizers>=0.13.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Image Processing
Pillow>=9.5.0

# Web App
streamlit>=1.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
